storeを統一
