+++ frontend/src/hooks/useMailStore.ts
@@ -1,9 +1,10 @@
 import { create } from "zustand";
 import { Mail, MailTable, PostMail } from "@/types/mail";
 import { mockMailList } from "@/fixtures/mail-list";
-// import { endpoints } from "@/utils/apiUrls";
-// import { apiClient } from "@/domain/api/apiClient";
-// import { MailGroup } from "@/types/mail";
+import { MailGroup } from "@/types/mail";
+import { User } from "@/types/user";
+import { apiClient } from "@/domain/api/apiClient";
+import { endpoints } from "@/utils/apiUrls";
 
 type MailStore = {
   postMail: PostMail | null;
@@ -13,13 +14,27 @@ type MailStore = {
   getMails: () => void;
   getMailTableList: () => void;
   sendMail: (mailGroupId: number) => void;
+  mailGroupList: MailGroup[];
+  setMailGroupList: (mailGroupList: MailGroup[]) => void;
+  getMailGroupList: () => void;
+  createMailGroup: (selectedUsers: User[]) => void;
+  groupTitle: string;
+  setGroupTitle: (groupTitle: string) => void;
+  groupNote: string;
+  setGroupNote: (groupNote: string) => void;
 };
 
-export const useMailStore = create<MailStore>((set) => ({
+export const useMailStore = create<MailStore>((set, get) => ({
   postMail: { title: "", message: "", group_id: 0 },
   setPostMail: (postMail: PostMail) => set({ postMail }),
   allMailList: [] as Mail[],
   allMailTableList: [] as MailTable[],
+  mailGroupList: [] as MailGroup[],
+  setMailGroupList: (mailGroupList: MailGroup[]) => set({ mailGroupList }),
+  groupTitle: "",
+  setGroupTitle: (groupTitle: string) => set({ groupTitle }),
+  groupNote: "",
+  setGroupNote: (groupNote: string) => set({ groupNote }),
 
   /**
    * メール一覧を取得する
@@ -61,5 +76,35 @@ export const useMailStore = create<MailStore>((set) => ({
       message: postMail.message,
     };
     console.log(mail);
+  },
+
+  /**
+   * メールグループ一覧を取得する
+   */
+  getMailGroupList: async (): Promise<void> => {
+    const response = await apiClient<MailGroup[]>({
+      url: endpoints.get.mailGroupList,
+      method: "GET",
+    });
+    console.log(response);
+    set({ mailGroupList: response });
+  },
+
+  /**
+   * メールグループを作成する
+   * TODO APIと連携する
+   * @param selectedUsers 選択されたユーザー
+   */
+  createMailGroup: (selectedUsers: User[]) => {
+    const { groupTitle, groupNote } = get();
+    const postData = {
+      group_title: groupTitle,
+      note: groupNote,
+      records: selectedUsers.map((user) => ({
+        recipient_user: user.id,
+        recipient_user_name: user.name,
+      })),
+    };
+    console.log(postData);
   },
 }));
